Change API to a builder interface.
The previous procedural interface not only exposed various details about the
implementation, but made it hard to expand. The aim of this builder interface is
to keep all the functionality that users will require but package it in a
slightly more abstract way.

This rewrite is the minimum possible rewrite into a builder style. Code along
the lines of:

  let opts = &[
    optopt("o", "", "set output file name", "NAME"),
    optflag("h", "help", "print this help menu")
  ];
  match getopts(args.tail(), opts) { ... }

can then be rewritten to:

  match Options::new()
                .add_optopt("o", "", "set output file name", "NAME")
                .add_optflag("h", "help", "print this help menu")
                .parse_freely(args.tail())
                { ... }