Fix formatting inconsistencies in documentation.
Several doc comments did not use a period to end a sentence. Since this was the
minority case, I added the periods. Also, the opt_default method did not use the
enum constructor None, but instead used the word none.

Additionally, a comment in the format_option function had two typos. They are
now fixed.