Parse options without names vector
The only time the names vector has more than one `Name` in it is when parsing contiguous short options (`-abc`). We would continue putting `Name`s into the vector until we hit a short option that takes an argument. At that point, we break, so we can process the argument.

So there's no point to reprocess the head of that vector, we already know they don't take arguments. Only the last one matters.