Parse optional short options like gnu getopt
Similar to https://github.com/rust-lang-nursery/getopts/issues/49, this changes the parsing of short options to require no space between the option and argument. Now, `-aSomething` parse as an option with argument, and `-a Something` will parse as a present flag with a free non-option.

This is a breaking change.