Add helpflag() to `Options`
When we add a flag with `reqopt()`, the mandatory flag is always
required. This behavior doesn't work well when we want to add a `help`
flag, which should work without the required flag is not specified like
`my_command --help`.

So, let's add `helpflag()` method, which add an optional flag "help"
which will allow `parse()` to skip checking the existence of mandatory
flags.