Document and start testing the MSRV
Check the MSRV in CI and document it as an extremely low 1.49 (the
oldest version I could test with locally, it's probably actually older
than that). This is intended so we can do one final release with
`rust-version` documented, then increase it in a follow up to 1.66 which
will be needed for `unicode-width` 0.2.